Memperbaiki desain web page
