+++ StackExchange Client/nbproject/jaxws-build.xml
@@ -42,5 +42,21 @@
         <delete dir="${build.generated.sources.dir}/jax-ws/UserWS"/>
         <delete dir="${build.generated.dir}/jax-wsCache/UserWS"/>
     </target>
-    <target name="wsimport-client-generate" depends="wsimport-client-UserWS"/>
+    <target name="wsimport-client-QuestionWS" depends="wsimport-init">
+        <mkdir dir="${build.generated.dir}/jax-wsCache/QuestionWS"/>
+        <wsimport sourcedestdir="${build.generated.dir}/jax-wsCache/QuestionWS" destdir="${build.generated.dir}/jax-wsCache/QuestionWS" wsdl="${basedir}/${conf-dir}xml-resources/web-service-references/QuestionWS/wsdl/localhost_8081/StackExchange_WS/QuestionWS.wsdl" catalog="catalog.xml" encoding="${source.encoding}" extension="true" verbose="true" fork="false" wsdlLocation="http://localhost:8081/StackExchange_WS/QuestionWS?wsdl" xnocompile="true" xendorsed="true" package="QuestionWS">
+            <depends file="${basedir}/${conf-dir}xml-resources/web-service-references/QuestionWS/wsdl/localhost_8081/StackExchange_WS/QuestionWS.wsdl"/>
+            <produces dir="${build.generated.dir}/jax-wsCache/QuestionWS"/>
+        </wsimport>
+        <copy todir="${build.generated.sources.dir}/jax-ws">
+            <fileset dir="${build.generated.dir}/jax-wsCache/QuestionWS">
+                <include name="**/*.java"/>
+            </fileset>
+        </copy>
+    </target>
+    <target name="wsimport-client-clean-QuestionWS" depends="-init-project">
+        <delete dir="${build.generated.sources.dir}/jax-ws/QuestionWS"/>
+        <delete dir="${build.generated.dir}/jax-wsCache/QuestionWS"/>
+    </target>
+    <target name="wsimport-client-generate" depends="wsimport-client-UserWS, wsimport-client-QuestionWS"/>
 </project>
