+++ StackExchange Client/web/WEB-INF/glassfish-web.xml
@@ -2,9 +2,6 @@
 <!DOCTYPE glassfish-web-app PUBLIC "-//GlassFish.org//DTD GlassFish Application Server 3.1 Servlet 3.0//EN" "http://glassfish.org/dtds/glassfish-web-app_3_0-1.dtd">
 <glassfish-web-app error-url="">
   <context-root>/StackExchange_Client</context-root>
-  <servlet>
-    <servlet-name>Test</servlet-name>
-  </servlet>
   <jsp-config>
     <property name="keepgenerated" value="true">
       <description>Keep a copy of the generated servlet class' java code.</description>
