+++ StackExchange Client/web/WEB-INF/jax-ws-catalog.xml
@@ -4,4 +4,6 @@
     <system systemId="http://localhost:8081/StackExchange_WS/AnswerWS?xsd=1" uri="wsdl/localhost_8081/StackExchange_WS/AnswerWS.xsd_1.xsd"/>
     <system systemId="http://localhost:8081/StackExchange_WS/UserWS?wsdl" uri="wsdl/localhost_8081/StackExchange_WS/UserWS.wsdl"/>
     <system systemId="http://localhost:8081/StackExchange_WS/UserWS?xsd=1" uri="wsdl/localhost_8081/StackExchange_WS/UserWS.xsd_1.xsd"/>
+    <system systemId="http://localhost:8081/StackExchange_WS/QuestionWS?wsdl" uri="wsdl/localhost_8081/StackExchange_WS/QuestionWS.wsdl"/>
+    <system systemId="http://localhost:8081/StackExchange_WS/QuestionWS?xsd=1" uri="wsdl/localhost_8081/StackExchange_WS/QuestionWS.xsd_1.xsd"/>
 </catalog>
