+++ StackExchange Client/web/WEB-INF/web.xml
@@ -1,12 +1,20 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <web-app version="3.1" xmlns="http://xmlns.jcp.org/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/javaee http://xmlns.jcp.org/xml/ns/javaee/web-app_3_1.xsd">
     <servlet>
-        <servlet-name>UserServlet</servlet-name>
-        <servlet-class>User.UserServlet</servlet-class>
+        <servlet-name>RegisterServlet</servlet-name>
+        <servlet-class>User.RegisterServlet</servlet-class>
+    </servlet>
+    <servlet>
+        <servlet-name>QuestionListServlet</servlet-name>
+        <servlet-class>Question.QuestionListServlet</servlet-class>
     </servlet>
     <servlet-mapping>
-        <servlet-name>UserServlet</servlet-name>
-        <url-pattern>/UserServlet</url-pattern>
+        <servlet-name>RegisterServlet</servlet-name>
+        <url-pattern>/RegisterServlet</url-pattern>
+    </servlet-mapping>
+    <servlet-mapping>
+        <servlet-name>QuestionListServlet</servlet-name>
+        <url-pattern>/QuestionListServlet</url-pattern>
     </servlet-mapping>
     <session-config>
         <session-timeout>
