+++ StackExchange Client/web/WEB-INF/wsdl/localhost_8081/StackExchange_WS/QuestionWS.wsdl
@@ -1,0 +1,131 @@
+<?xml version='1.0' encoding='UTF-8'?><!-- Published by JAX-WS RI (http://jax-ws.java.net). RI's version is Metro/2.3.1-b419 (branches/2.3.1.x-7937; 2014-08-04T08:11:03+0000) JAXWS-RI/2.2.10-b140803.1500 JAXWS-API/2.2.11 JAXB-RI/2.2.10-b140802.1033 JAXB-API/2.2.12-b140109.1041 svn-revision#unknown. --><!-- Generated by JAX-WS RI (http://jax-ws.java.net). RI's version is Metro/2.3.1-b419 (branches/2.3.1.x-7937; 2014-08-04T08:11:03+0000) JAXWS-RI/2.2.10-b140803.1500 JAXWS-API/2.2.11 JAXB-RI/2.2.10-b140802.1033 JAXB-API/2.2.12-b140109.1041 svn-revision#unknown. --><definitions xmlns:wsu="http://docs.oasis-open.org/wss/2004/01/oasis-200401-wss-wssecurity-utility-1.0.xsd" xmlns:wsp="http://www.w3.org/ns/ws-policy" xmlns:wsp1_2="http://schemas.xmlsoap.org/ws/2004/09/policy" xmlns:wsam="http://www.w3.org/2007/05/addressing/metadata" xmlns:soap="http://schemas.xmlsoap.org/wsdl/soap/" xmlns:tns="http://QuestionModel/" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns="http://schemas.xmlsoap.org/wsdl/" targetNamespace="http://QuestionModel/" name="QuestionWS">
+<types>
+<xsd:schema>
+<xsd:import namespace="http://QuestionModel/" schemaLocation="http://localhost:8081/StackExchange_WS/QuestionWS?xsd=1"/>
+</xsd:schema>
+</types>
+<message name="editQuestion">
+<part name="parameters" element="tns:editQuestion"/>
+</message>
+<message name="editQuestionResponse">
+<part name="parameters" element="tns:editQuestionResponse"/>
+</message>
+<message name="voteQuestion">
+<part name="parameters" element="tns:voteQuestion"/>
+</message>
+<message name="voteQuestionResponse">
+<part name="parameters" element="tns:voteQuestionResponse"/>
+</message>
+<message name="getQuestionByQID">
+<part name="parameters" element="tns:getQuestionByQID"/>
+</message>
+<message name="getQuestionByQIDResponse">
+<part name="parameters" element="tns:getQuestionByQIDResponse"/>
+</message>
+<message name="deleteQuestion">
+<part name="parameters" element="tns:deleteQuestion"/>
+</message>
+<message name="deleteQuestionResponse">
+<part name="parameters" element="tns:deleteQuestionResponse"/>
+</message>
+<message name="getAllQuestion">
+<part name="parameters" element="tns:getAllQuestion"/>
+</message>
+<message name="getAllQuestionResponse">
+<part name="parameters" element="tns:getAllQuestionResponse"/>
+</message>
+<message name="createQuestion">
+<part name="parameters" element="tns:createQuestion"/>
+</message>
+<message name="createQuestionResponse">
+<part name="parameters" element="tns:createQuestionResponse"/>
+</message>
+<portType name="QuestionWS">
+<operation name="editQuestion">
+<input wsam:Action="http://QuestionModel/QuestionWS/editQuestionRequest" message="tns:editQuestion"/>
+<output wsam:Action="http://QuestionModel/QuestionWS/editQuestionResponse" message="tns:editQuestionResponse"/>
+</operation>
+<operation name="voteQuestion">
+<input wsam:Action="http://QuestionModel/QuestionWS/voteQuestionRequest" message="tns:voteQuestion"/>
+<output wsam:Action="http://QuestionModel/QuestionWS/voteQuestionResponse" message="tns:voteQuestionResponse"/>
+</operation>
+<operation name="getQuestionByQID">
+<input wsam:Action="http://QuestionModel/QuestionWS/getQuestionByQIDRequest" message="tns:getQuestionByQID"/>
+<output wsam:Action="http://QuestionModel/QuestionWS/getQuestionByQIDResponse" message="tns:getQuestionByQIDResponse"/>
+</operation>
+<operation name="deleteQuestion">
+<input wsam:Action="http://QuestionModel/QuestionWS/deleteQuestionRequest" message="tns:deleteQuestion"/>
+<output wsam:Action="http://QuestionModel/QuestionWS/deleteQuestionResponse" message="tns:deleteQuestionResponse"/>
+</operation>
+<operation name="getAllQuestion">
+<input wsam:Action="http://QuestionModel/QuestionWS/getAllQuestionRequest" message="tns:getAllQuestion"/>
+<output wsam:Action="http://QuestionModel/QuestionWS/getAllQuestionResponse" message="tns:getAllQuestionResponse"/>
+</operation>
+<operation name="createQuestion">
+<input wsam:Action="http://QuestionModel/QuestionWS/createQuestionRequest" message="tns:createQuestion"/>
+<output wsam:Action="http://QuestionModel/QuestionWS/createQuestionResponse" message="tns:createQuestionResponse"/>
+</operation>
+</portType>
+<binding name="QuestionWSPortBinding" type="tns:QuestionWS">
+<soap:binding transport="http://schemas.xmlsoap.org/soap/http" style="document"/>
+<operation name="editQuestion">
+<soap:operation soapAction=""/>
+<input>
+<soap:body use="literal"/>
+</input>
+<output>
+<soap:body use="literal"/>
+</output>
+</operation>
+<operation name="voteQuestion">
+<soap:operation soapAction=""/>
+<input>
+<soap:body use="literal"/>
+</input>
+<output>
+<soap:body use="literal"/>
+</output>
+</operation>
+<operation name="getQuestionByQID">
+<soap:operation soapAction=""/>
+<input>
+<soap:body use="literal"/>
+</input>
+<output>
+<soap:body use="literal"/>
+</output>
+</operation>
+<operation name="deleteQuestion">
+<soap:operation soapAction=""/>
+<input>
+<soap:body use="literal"/>
+</input>
+<output>
+<soap:body use="literal"/>
+</output>
+</operation>
+<operation name="getAllQuestion">
+<soap:operation soapAction=""/>
+<input>
+<soap:body use="literal"/>
+</input>
+<output>
+<soap:body use="literal"/>
+</output>
+</operation>
+<operation name="createQuestion">
+<soap:operation soapAction=""/>
+<input>
+<soap:body use="literal"/>
+</input>
+<output>
+<soap:body use="literal"/>
+</output>
+</operation>
+</binding>
+<service name="QuestionWS">
+<port name="QuestionWSPort" binding="tns:QuestionWSPortBinding">
+<soap:address location="http://localhost:8081/StackExchange_WS/QuestionWS"/>
+</port>
+</service>
+</definitions>
