+++ StackExchange Client/web/edit.jsp
@@ -4,19 +4,29 @@
     <head>
         <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
         <title>Simple StackExchange</title>
-        <link rel="stylesheet" href="css/main.css">
+        <link rel="stylesheet" href="main.css">
     </head>
     <body>
-        <h1>
-            <a href="index.jsp">Simple StackExchange</a>
-        </h1>
-        <br>
-        <h2>What's your question?</h2>
-        <hr>
-        <form name="editForm" action="" method="post" onsubmit="">
-            <input name="topic" class="text" type="text" placeholder="Question Topic"><br>
-            <textarea name="content" placeholder="Content"></textarea>
-            <input class="button" type="submit" value="Post"><br>
-        </form>
+        <div class="navbar-up">
+            <a href="index.jsp">
+                <h1 class="white">Simple StackExchange</h1>
+            </a>
+        </div>
+        <div class="navbar-down">
+            <b class="navbar-down">
+                <a class="white" href="register.jsp">Register</a> |
+                <a class="white" href="login.jsp">Login</a>
+            </b>
+        </div>
+        <div class="main">
+            <br>
+            <h2>Edit your question</h2>
+            <hr>
+            <form name="editForm" action="" method="post" onsubmit="">
+                <input name="topic" class="text" type="text" placeholder="Question Topic"><br>
+                <textarea name="content" placeholder="Content"></textarea>
+                <input class="button" type="submit" value="Post"><br>
+            </form>
+        </div>
     </body>
 </html>
