+++ StackExchange Client/web/login.jsp
@@ -4,14 +4,29 @@
     <head>
         <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
         <title>Simple StackExchange</title>
-        <link rel="stylesheet" href="css/main.css">
+        <link rel="stylesheet" href="main.css">
     </head>
     <body>
-        <h1>
-            <a href="index.jsp">Simple StackExchange</a>
-        </h1>
-        <input name="email" type="text" placeholder="Email"><br>
-        <input name="password" type="password" placeholder="Password"><br>
-        <input type="submit" value="Login">
+        <div class="navbar-up">
+            <a href="index.jsp">
+                <h1 class="white">Simple StackExchange</h1>
+            </a>
+        </div>
+        <div class="navbar-down">
+            <b class="navbar-down">
+                <a class="white" href="register.jsp">Register</a> |
+                <a class="white" href="login.jsp">Login</a>
+            </b>
+        </div>
+        <div class="main">
+            <br>
+            <h2>Log in to your account</h2>
+            <hr>
+            <form>
+                <input class="text" name="email" type="text" placeholder="Email"><br>
+                <input class="text" name="password" type="password" placeholder="Password"><br>
+                <input type="submit" value="Login">
+            </form>
+        </div>
     </body>
 </html>
